comfyui-experiment/pyproject.toml
[project]
name = "comfyui-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.0",
    "comfy-script>=0.5.1",
    "jinja2>=3.1.6",
    "nest-asyncio>=1.6.0",
    "pillow>=11.3.0",
    "pytest>=8.4.2",
    "requests>=2.32.5",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "websocket-client>=1.9.0",
]

[tool.pytest.ini_options]
addopts = "-s -v"
